requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.2.2
cffi==1.17.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-autocomplete-light==3.11.0
django-cors-headers==4.3.1
django-import-export==3.2.0
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
heyoo==0.1.0
hiredis==3.0.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
MarkupPy==1.14
msgpack==1.1.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.1.1
reportlab==4.0.4
requests==2.31.0
requests-toolbelt==1.0.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
Twisted==24.7.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.1
whatsapp-cloud-api==0.0.4
xlrd==2.0.1
xlwt==1.3.0
zope.interface==7.1.0
